Cleanup of coding style and number casting

+++ ProjectFolio/CCDetailViewController.m
@@ -512,8 +512,8 @@
 void (^completionHandler)(UIPrintInteractionController *, BOOL, NSError *) =
 ^(UIPrintInteractionController *pic, BOOL completed, NSError *error) {
     if (!completed && error)
-        NSLog(@"FAILED! due to error in domain %@ with error code %u",
-              error.domain, error.code);
+        NSLog(@"FAILED! due to error in domain %@ with error code %ld",
+              error.domain, (long)error.code);
 };
 
 -(void)releaseLogger{
